float to double + minimalistic astyle

+++ src/cata_utility.h
@@ -49,7 +49,7 @@ const char *velocity_units( const units_type vel_units );
 const char *weight_units();
 double convert_velocity( int velocity, const units_type vel_units );
 double convert_weight( int weight );
-double temp_to_celsius( float fahrenheit );
+double temp_to_celsius( double fahrenheit );
 
 /**
 * @brief Class used to access a list as if it were circular.
