+++ src/weather.cpp
@@ -575,18 +575,18 @@ std::string weather_forecast( point const &abs_sm_pos )
 /**
  * Print temperature (and convert to celsius if celsius display is enabled.)
  */
-std::string print_temperature(float fahrenheit, int decimals)
+std::string print_temperature( double fahrenheit, int decimals )
 {
     std::ostringstream ret;
-    ret.precision(decimals);
+    ret.precision( decimals );
     ret << std::fixed;
 
     if(OPTIONS["USE_CELSIUS"] == "celsius") {
         ret << temp_to_celsius( fahrenheit );
-        return rmp_format(_("<Celsius>%sC"), ret.str().c_str());
+        return rmp_format( _( "<Celsius>%sC" ), ret.str().c_str() );
     } else {
         ret << fahrenheit;
-        return rmp_format(_("<Fahrenheit>%sF"), ret.str().c_str());
+        return rmp_format( _( "<Fahrenheit>%sF" ), ret.str().c_str() );
     }
 
 }
@@ -594,7 +594,7 @@ std::string print_temperature(float fahrenheit, int decimals)
 /**
  * Print relative humidity (no conversions.)
  */
-std::string print_humidity( float humidity, int decimals )
+std::string print_humidity( double humidity, int decimals )
 {
     std::ostringstream ret;
     ret.precision( decimals );
@@ -607,18 +607,18 @@ std::string print_humidity( float humidity, int decimals )
 /**
  * Print pressure (no conversions.)
  */
-std::string print_pressure(float pressure, int decimals)
+std::string print_pressure( double pressure, int decimals )
 {
     std::ostringstream ret;
-    ret.precision(decimals);
+    ret.precision( decimals );
     ret << std::fixed;
 
-    ret << pressure/10;
-    return rmp_format(_("%s kPa"), ret.str().c_str());
+    ret << pressure / 10;
+    return rmp_format( _( "%s kPa" ), ret.str().c_str() );
 }
 
 
-int get_local_windchill(double temperature, double humidity, double windpower)
+int get_local_windchill( double temperature, double humidity, double windpower )
 {
     double tmptemp = temperature;
     double tmpwind = windpower;
@@ -651,7 +651,7 @@ int get_local_windchill(double temperature, double humidity, double windpower)
     return windchill;
 }
 
-int get_local_humidity(double humidity, weather_type weather, bool sheltered)
+int get_local_humidity( double humidity, weather_type weather, bool sheltered )
 {
     int tmphumidity = humidity;
     if( sheltered ) {
